--- a/src/Context/Cache.hs
+++ b/src/Context/Cache.hs
@@ -22,7 +22,7 @@ saveCache :: Source.Source -> Cache.Cache -> App ()
 saveCache source cache = do
   cachePath <- Path.getSourceCachePath source
   ensureDir $ parent cachePath
-  liftIO $ encodeFile (toFilePath cachePath) cache
+  liftIO $ encodeFile (toFilePath cachePath) $ Cache.compress cache
 
 loadCache :: Source.Source -> App (Maybe Cache.Cache)
 loadCache source = do