For better compilation speed, this PR:

- parallelizes the build process to some extent,
- drops location information when saving cache, and
- optimizes the internal representation of definite descriptions.

---

```
before:
❯ hyperfine "neut build"
Benchmark 1: neut build
  Time (mean ± σ):      1.323 s ±  0.023 s    [User: 1.932 s, System: 0.313 s]
  Range (min … max):    1.277 s …  1.351 s    10 runs

after:
❯ hyperfine "neut build"
Benchmark 1: neut build
  Time (mean ± σ):      83.1 ms ±   3.9 ms    [User: 125.5 ms, System: 26.5 ms]
  Range (min … max):    76.6 ms …  94.7 ms    36 runs
```